add optional wiping, gpg signature verification, rslave mounting to stage3 mount script

+++ src/ansible/scripts/mount-stage3.sh
@@ -15,34 +15,59 @@ cryptsetup close root || true
 swapoff /dev/mapper/swap || true
 cryptsetup close swap || true
 
-# Format filesystems.
-yes | mke2fs "$BOOT_DEV"
-cryptsetup -d /dev/urandom create swap "$SWAP_DEV"
-mkswap /dev/mapper/swap
-swapon /dev/mapper/swap
-cryptsetup --verbose --batch-mode luksFormat "$ROOT_DEV" <<<"$LUKS_PASS"
+# Set up LUKS.
+if [ "$WIPE" = yes ]; then
+    cryptsetup --verbose --batch-mode luksFormat "$ROOT_DEV" <<<"$LUKS_PASS"
+fi
 cryptsetup --verbose --batch-mode open --type luks "$ROOT_DEV" root <<<"$LUKS_PASS"
-mkfs.btrfs /dev/mapper/root
+cryptsetup -d /dev/urandom create swap "$SWAP_DEV"
 
-# Mount root filesystem.
+# Set up filesystems.
+if [ "$WIPE" = yes ]; then
+    mkfs.btrfs /dev/mapper/root
+    mkswap /dev/mapper/swap
+    yes | mke2fs "$BOOT_DEV"
+fi
 mkdir /mnt/root
 mount -o noatime /dev/mapper/root /mnt/root
-btrfs su create /mnt/root/root
-btrfs su create /mnt/root/portage
 
-# Extract stage3.
-cd /tmp
-test -e "${STAGE3_URL##*/}" || wget "$STAGE3_URL"
-tar xjpf "${STAGE3_URL##*/}" -C /mnt/root/root
+# Populate root filesystem.
+if [ "$WIPE" = yes ]; then
+    btrfs su create /mnt/root/root
+    btrfs su create /mnt/root/portage
 
-# Create snapshot post-stage3.
-mkdir /mnt/root/.snapshots
-sdate() { date -u +"%Y%m%dT%H%M%SZ"; }
-btrfs su snapshot -r /mnt/root/root /mnt/root/.snapshots/root@$(sdate)
+    # Extract stage3.
+    cd /tmp
+    if ! [ -e "${STAGE3_URL##*/}" ]; then
+        wget --no-check-certificate "$STAGE3_URL"
+        wget --no-check-certificate "$STAGE3_URL.DIGESTS.asc"
+    fi
+    gpg --verify "${STAGE3_URL##*/}.DIGESTS.asc"
+    grep $(sha512sum "${STAGE3_URL##*/}") <(gpg --decrypt "${STAGE3_URL##*/}.DIGESTS.asc")
+    tar xjpf "${STAGE3_URL##*/}" --xattrs -C /mnt/root/root
 
-# Set up /mnt/gentoo.
-mkdir /mnt/root/root/usr/portage
+    # Create snapshot post-stage3.
+    mkdir /mnt/root/.snapshots
+    sdate() { date -u +"%Y%m%dT%H%M%SZ"; }
+    btrfs su snapshot -r /mnt/root/root /mnt/root/.snapshots/root@$(sdate)
+
+    mkdir /mnt/root/root/usr/portage
+fi
+
+# Set up runtime.
+swapon /dev/mapper/swap
 mount --bind /mnt/root/portage /mnt/root/root/usr/portage
 mount -o noatime "$BOOT_DEV" /mnt/root/root/boot
+mount -t proc proc /mnt/gentoo/proc
+mount --rbind /sys /mnt/gentoo/sys
+mount --make-rslave /mnt/gentoo/sys
+if [ -L /dev/shm ]; then
+    mount -t tmpfs -o nosuid,nodev,noexec shm /dev/shm
+    chmod 1777 /dev/shm
+fi
+mount --rbind /dev /mnt/gentoo/dev
+mount --make-rslave /mnt/gentoo/dev
+mount --rbind /tmp /mnt/root/root/tmp
+cp -Lnv /etc/resolv.conf /mnt/root/root/etc/resolv.conf
 mkdir /mnt/gentoo
 mount --rbind /mnt/root/root /mnt/gentoo
